feat: updateAttributes 함수 공통화
